requirements.txt
ast
cv2==3.4.2
imgaug==0.2.6
imutils==0.5.1
json==2.0.9
matplotlib==2.2.2
numpy==1.15.2
pandas==0.23.0
pickle
PIL==5.1.0
shapely==1.6.4
urllib
python-dotenv==0.15.0
